requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asciitree==0.3.3
asttokens==2.4.1
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
azure-core==1.30.2
azure-storage-blob==12.20.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==19.10b0
boto3==1.34.120
botocore==1.34.120
braceexpand==0.1.7
cdifflib==1.2.6
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.0.2
clip==0.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
cytoolz==0.12.3
datasets==2.19.2
debugpy==1.8.1
decorator==4.4.2
diffusers==0.28.2
dill==0.3.8
Distance==0.1.3
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.0
einops-exts==0.0.4
email_validator==2.1.1
environs==9.5.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.19
fasttext==0.9.2
filelock==3.14.0
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.60.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
hydra-core==1.3.2
idna==3.7
ijson==3.3.0
imageio==2.34.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==7.1.0
inflect==7.2.1
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiwer==2.5.2
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.22.0
lhotse==1.24.0
librosa==0.10.2.post1
lightning-utilities==0.11.2
lilcom==1.7
llvmlite==0.42.0
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
megatron_core==0.5.0
milvus-lite==2.4.7
minio==7.2.7
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nemo_text_processing==1.0.2
nemo_toolkit==1.23.0
nerfacc==0.5.3
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.16.1
open-clip-torch==2.24.0
OpenCC==1.1.6
orjson==3.10.4
packaging==24.1
pandas==2.2.2
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
plac==1.4.3
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.8.2
proglog==0.1.10
progress==1.6
prompt_toolkit==3.0.46
protobuf==5.27.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.12.0
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
PyMCubes==0.1.4
pymilvus==2.3.7
pynini==2.1.5
pyparsing==3.1.2
pypinyin==0.51.0
pypinyin-dict==0.8.0
PySocks==1.7.1
pytest==8.2.2
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==2.13.7
regex==2024.5.15
requests==2.32.3
resampy==0.4.3
rich==13.7.1
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
sentencepiece==0.2.0
sentry-sdk==2.5.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.5.0
soxr==0.3.7
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.2
texterrors==0.4.4
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchdiffeq==0.2.4
torchmetrics==1.4.0.post0
torchsde==0.2.6
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.41.2
trimesh==4.4.0
triton==2.3.1
typed-ast==1.5.5
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.1.62
websockets==12.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
zarr==2.18.2
zipp==3.19.2
